Refactor admin forms into a separate module
