+++ concordia/admin/forms.py
@@ -1,0 +1,14 @@
+from django import forms
+
+
+class AdminItemImportForm(forms.Form):
+    import_url = forms.URLField(
+        required=True, label="URL of the item/collection/search page to import"
+    )
+
+
+class AdminProjectBulkImportForm(forms.Form):
+    spreadsheet_file = forms.FileField(
+        required=True,
+        label="Spreadsheet containing the campaigns, projects, and items to import",
+    )
