+++ concordia/forms.py
@@ -59,16 +59,3 @@ class ContactUsForm(forms.Form):
     story = forms.CharField(
         label="Let us know how we can help:", required=True, widget=forms.Textarea
     )
-
-
-class AdminItemImportForm(forms.Form):
-    import_url = forms.URLField(
-        required=True, label="URL of the item/collection/search page to import"
-    )
-
-
-class AdminProjectBulkImportForm(forms.Form):
-    spreadsheet_file = forms.FileField(
-        required=True,
-        label="Spreadsheet containing the campaigns, projects, and items to import",
-    )
